07_pyrogrammod/pyproject.toml
[project]
name = "07-pyrogrammod"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "pyrogrammod>=2.2.4",
    "tgcrypto>=1.2.4"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"

